requirements.txt
Pillow == 8.3.2
pystray == 0.17.3